change grid to tabl
have changed grid to table because it allows us to have tables within a table which is needed i think? so that its easier to add those sticky notes that should be sticky vertically. let me know if you can think of a better way of doing this like going back to the grid way of it